[backend] Split backend.tsx file into smaller parts based on functionality
