+++ games/memory/MemoryStart.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import { NativeStackScreenProps } from "@react-navigation/native-stack";
 import { MemoryStackParamList } from "./MemoryStack";
-import { callTranslations } from "../../backend";
+import { callTranslations } from "../../backend/WordSetsBackend";
 import { useAppStore } from "../../state";
 import { makeCards } from "./MemoryCard";
 import SinglePlayerStartScreen from "../common_singleplayer/SingleplayerStart";
