+++ screens/Login.tsx
@@ -10,7 +10,7 @@ import {
 import React, { useState } from "react";
 import authStyles from "../styles/AuthStyles";
 import mainStyles from "../styles/MainStyles";
-import { callLogin } from "../backend";
+import { callLogin } from "../backend/AuthBackend";
 import { NativeStackScreenProps } from "@react-navigation/native-stack";
 import Toast from "react-native-toast-message";
 import { AuthStackParamList } from "../navgation/AuthStack";
