+++ screens/Register.tsx
@@ -10,7 +10,7 @@ import {
 import React, { useState } from "react";
 import authStyles from "../styles/AuthStyles";
 import mainStyles from "../styles/MainStyles";
-import { callLogin, callRegister } from "../backend";
+import { callLogin, callRegister } from "../backend/AuthBackend";
 import { NativeStackScreenProps } from "@react-navigation/native-stack";
 import { AuthStackParamList } from "../navgation/AuthStack";
 import Toast from "react-native-toast-message";
